+++ .idea/.gitignore
@@ -1,3 +1,6 @@
 # Default ignored files
 /shelf/
 /workspace.xml
+.idea
+/out/
+
