+++ src/com/javarush/ava111/Service/PersonService.java
@@ -1,4 +1,9 @@
-package com.javarush.ava111;
+package com.javarush.ava111.Service;
+
+import com.javarush.ava111.Dto.CreateUpdateDto;
+import com.javarush.ava111.Dto.DeleteDto;
+import com.javarush.ava111.Dto.ReadDto;
+import com.javarush.ava111.Person;
 
 import java.text.ParseException;
 import java.util.HashMap;
