--- src/com/javarush/ava111/Uitl.java
@@ -1,9 +1,0 @@
-package com.javarush.ava111;
-
-import java.text.SimpleDateFormat;
-
-import static com.javarush.ava111.ApplicationSettings.DATE_PATTERN;
-
-public class Uitl {
-    public static final SimpleDateFormat formatter = new SimpleDateFormat(DATE_PATTERN);
-}
